Incremental Mode Pre-Release

+++ src/main/java/dev/rgbmc/ferrum/api/Backup.java
@@ -1,11 +1,16 @@
 package dev.rgbmc.ferrum.api;
 
+import com.google.gson.Gson;
 import dev.rgbmc.ferrum.api.handlers.AbstractHandler;
 import dev.rgbmc.ferrum.api.handlers.SimpleHandler;
+import dev.rgbmc.ferrum.api.objects.IncrementalIndex;
 import dev.rgbmc.ferrum.api.objects.ResultInfo;
-import dev.rgbmc.ferrum.api.utils.DiffUtils;
+import dev.rgbmc.ferrum.api.utils.CRCUtils;
+import dev.rgbmc.ferrum.api.utils.IncrementalUtils;
 import dev.rgbmc.ferrum.api.utils.LogoUtils;
+import net.lingala.zip4j.ZipFile;
 import net.lingala.zip4j.io.outputstream.ZipOutputStream;
+import net.lingala.zip4j.model.FileHeader;
 import net.lingala.zip4j.model.ZipParameters;
 import net.lingala.zip4j.model.enums.AesKeyStrength;
 import net.lingala.zip4j.model.enums.CompressionLevel;
@@ -18,21 +23,33 @@ import java.io.BufferedInputStream;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.StandardOpenOption;
 import java.text.SimpleDateFormat;
 import java.util.*;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
+import java.util.concurrent.ThreadFactory;
 
 public class Backup {
 
     public final static Logger logger = LoggerFactory.getLogger("Ferrum");
     private static final SimpleDateFormat simpleDateFormat = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
-    private static final ExecutorService executorService = Executors.newFixedThreadPool(4);
+    private static final ExecutorService executorService = Executors.newFixedThreadPool(4, new ThreadFactory() {
+        private int count = 0;
+
+        @Override
+        public Thread newThread(Runnable r) {
+            count++;
+            return new Thread(r, "Ferrum-" + count);
+        }
+    });
     private final Path folder;
     private final List<AbstractHandler> handlers = new ArrayList<>(Collections.singletonList(new SimpleHandler()));
-    private final String zipPath;
+    private final Map<String, FileHeader> cachedFileHeader = new HashMap<>();
+    private String zipPath;
     private File file;
     private String password = null;
     private boolean encrypt = false;
@@ -63,39 +80,70 @@ public class Backup {
         executorService.submit(runnable);
     }
 
+    public static void shutdown() {
+        executorService.shutdownNow();
+    }
+
     public ResultInfo startBackup() {
         //System.out.println(zipPath);
         try {
             ZipParameters zipParameters = getZipParameters();
-            if (!incremental) {
+            if (!incremental && file.exists()) {
                 logger.warn("Zip File already exist, Ferrum will delete it when progress start");
                 file.delete();
             }
-            long count = Arrays.stream(file.getParentFile().listFiles()).filter(file1 -> file1.getName().endsWith(".zip")).count();
+            long count = Arrays.stream(file.getParentFile().listFiles()).filter(file1 -> file1.getName().endsWith(".full.zip")).count();
             ZipOutputStream zipOutputStream = initializeZipOutputStream(
                     file,
                     password != null,
                     (password == null ? new char[0] : password.toCharArray()),
                     incremental
             );
-            walkingFile(zipOutputStream, zipParameters, folder.toFile());
-            if (incremental) logger.info("Temp File has finished creating");
-            zipOutputStream.setComment(getComment());
-            zipOutputStream.close();
             if (incremental) {
                 if (count >= 1) {
-                    String originalFileName = file.getName();
-                    file = new File(file.getParentFile(), originalFileName + ".tmp");
-                    File diffFile = new File(file.getParentFile(), originalFileName + ".patch");
-                    FileOutputStream fileOutputStream = new FileOutputStream(diffFile);
-                    File zip = Arrays.stream(file.getParentFile().listFiles()).filter(file1 -> file1.getName().endsWith(".zip")).findFirst().get();
-                    DiffUtils.createDiffFile(zip, file, diffFile);
-                    file.delete();
-                    file = diffFile;
+                    File originZip = IncrementalUtils.findOriginZip(file.getParentFile());
+                    ZipFile originZipFile;
+                    if (encrypt) {
+                        originZipFile = new ZipFile(originZip, password.toCharArray());
+                    } else {
+                        originZipFile = new ZipFile(originZip);
+                    }
+                    IncrementalIndex index = new IncrementalIndex();
+                    for (FileHeader fileHeader : originZipFile.getFileHeaders()) {
+                        cachedFileHeader.put(fileHeader.getFileName(), fileHeader);
+                        Path realPath = folder.resolve(fileHeader.getFileName());
+                        File realFile = realPath.toFile();
+                        if (!realFile.exists()) {
+                            index.getDeletion().add(fileHeader.getFileName());
+                        } else if (fileHeader.getCrc() != CRCUtils.getCRC(realFile)) {
+                            index.getModification().add(fileHeader.getFileName());
+                            writeFileToZip(zipOutputStream, zipParameters, realFile);
+                        }
+                    }
+                    walkingFile(index, zipOutputStream, folder.toFile(), zipParameters);
+                    File tempFile = File.createTempFile("Ferrum_Incremental_", ".json");
+                    String json = new Gson().toJson(index);
+                    Files.write(tempFile.toPath(), json.getBytes(StandardCharsets.UTF_8),
+                            StandardOpenOption.CREATE,
+                            StandardOpenOption.WRITE,
+                            StandardOpenOption.TRUNCATE_EXISTING);
+                    writeFileToZip(zipOutputStream, getParameters(zipParameters, "ferrum_index.json"), tempFile, false);
+                    zipOutputStream.setComment(getComment() + "\n[Incremental Mode File]");
+                    zipOutputStream.close();
+                    tempFile.deleteOnExit();
+                    originZipFile.close();
+                    return new ResultInfo(this, file);
+                } else {
+                    zipPath = zipPath.replace(".zip", ".full.zip");
+                    file = new File(file.getParentFile(), file.getName().replace(".zip", ".full.zip"));
+                    logger.info("[Incremental Mode] Creating full-file zip for next time");
                 }
             }
+            walkingFile(zipOutputStream, zipParameters, folder.toFile());
+            zipOutputStream.setComment(getComment());
+            zipOutputStream.close();
         } catch (Exception e) {
-            e.printStackTrace();
+            e.printStackTrace(System.err);
             throw new RuntimeException(e);
         }
         handlers.forEach(handler -> {
@@ -120,6 +168,22 @@ public class Backup {
         }
     }
 
+    private void walkingFile(IncrementalIndex index, ZipOutputStream zipOutputStream, File file, ZipParameters zipParameters) {
+        Path walkingPath = file.toPath();
+        Path relativizedPath = folder.relativize(walkingPath);
+        //System.out.println(folder.relativize(file.toPath()));
+        if (relativizedPath.toString().contains(zipPath)) return;
+        if (ignores.stream().anyMatch(s -> relativizedPath.toString().startsWith(s) || relativizedPath.toString().endsWith(s)))
+            return;
+        if (file.isDirectory()) {
+            for (File children : file.listFiles()) {
+                walkingFile(index, zipOutputStream, children, zipParameters);
+            }
+        } else {
+            saveFile(index, zipOutputStream, file, zipParameters);
+        }
+    }
+
     private void saveFile(ZipOutputStream zipOutputStream, ZipParameters zipParameters, File file) {
         Path walkingPath = file.toPath();
         if (file.canWrite() && file.canRead()) {
@@ -128,15 +192,7 @@ public class Backup {
             if (ignores.stream().anyMatch(s -> relativizedPath.toString().startsWith(s) || relativizedPath.toString().endsWith(s)))
                 return;
             try {
-                zipOutputStream.putNextEntry(getParameters(zipParameters, file, folder));
-                byte[] buff = new byte[4096];
-                try (BufferedInputStream inputStream = new BufferedInputStream(Files.newInputStream(walkingPath), 4096)) {
-                    int readLen;
-                    while ((readLen = inputStream.read(buff)) != -1) {
-                        zipOutputStream.write(buff, 0, readLen);
-                    }
-                }
-                zipOutputStream.closeEntry();
+                writeFileToZip(zipOutputStream, zipParameters, file);
             } catch (Exception e) {
                 e.printStackTrace();
                 logger.error("[Skipped] Failed to zipping file " + relativizedPath + ", Error message: " + e.getMessage());
@@ -146,11 +202,44 @@ public class Backup {
         }
     }
 
-    private ZipOutputStream initializeZipOutputStream(File outputZipFile, boolean encrypt, char[] password, boolean incremental) throws IOException {
-        if (Arrays.stream(outputZipFile.getParentFile().listFiles()).anyMatch(file1 -> file1.getName().endsWith(".zip"))) {
-            outputZipFile = new File(outputZipFile.getParentFile(), outputZipFile.getName() + ".tmp");
-        }
+    private void writeFileToZip(ZipOutputStream zipOutputStream, ZipParameters zipParameters, File file) throws IOException {
+        writeFileToZip(zipOutputStream, zipParameters, file, true);
+    }
 
+    private void writeFileToZip(ZipOutputStream zipOutputStream, ZipParameters zipParameters, File file, boolean newParameters) throws IOException {
+        if (newParameters) zipParameters = getParameters(zipParameters, file, folder);
+        zipOutputStream.putNextEntry(zipParameters);
+        byte[] buff = new byte[4096];
+        try (BufferedInputStream inputStream = new BufferedInputStream(Files.newInputStream(file.toPath()), 4096)) {
+            int readLen;
+            while ((readLen = inputStream.read(buff)) != -1) {
+                zipOutputStream.write(buff, 0, readLen);
+            }
+        }
+        zipOutputStream.closeEntry();
+    }
+
+    private void saveFile(IncrementalIndex index, ZipOutputStream zipOutputStream, File file, ZipParameters zipParameters) {
+        Path walkingPath = file.toPath();
+        if (file.canWrite() && file.canRead()) {
+            Path relativizedPath = folder.relativize(walkingPath);
+            if (relativizedPath.toString().contains(zipPath)) return;
+            if (ignores.stream().anyMatch(s -> relativizedPath.toString().startsWith(s) || relativizedPath.toString().endsWith(s)))
+                return;
+            try {
+                if (cachedFileHeader.containsKey(relativizedPath.toString())) return;
+                writeFileToZip(zipOutputStream, zipParameters, file);
+                index.getCreation().add(relativizedPath.toString());
+            } catch (Exception e) {
+                e.printStackTrace();
+                logger.error("[Skipped] Failed to checking file " + relativizedPath + ", Error message: " + e.getMessage());
+            }
+        } else {
+            logger.warn("File " + folder.relativize(walkingPath) + " has been locked by other process or system, Ferrum skipped compress this file");
+        }
+    }
+
+    private ZipOutputStream initializeZipOutputStream(File outputZipFile, boolean encrypt, char[] password, boolean incremental) throws IOException {
         FileOutputStream fos = new FileOutputStream(outputZipFile);
 
         if (!encrypt) {
@@ -181,6 +270,12 @@ public class Backup {
         ZipParameters copy = new ZipParameters(zipParameters);
         Path relativizedPath = folder.relativize(file.toPath());
         copy.setFileNameInZip(relativizedPath.toString());
+        return copy;
+    }
+
+    private ZipParameters getParameters(ZipParameters zipParameters, String name) {
+        ZipParameters copy = new ZipParameters(zipParameters);
+        copy.setFileNameInZip(name);
         return copy;
     }
